[Breaking] More Ergonomic Workflow
This changes how this plugin is used. Instead of hard coding
api-extractor config in the repo its now required that you generate the
config using `docusuarus api-extractor:init`. This will produce a
config api-extractor.json file that will be used to configure the
extraction.

Documentation generation now uses a .api-extractor-meta file which is
just a compacted version of `Entry`s from walk-sync. For now we will use
this file to persist the mtimes, paths, and sizes of files in source.
When the documentation generation runs this file will be used to
determine if the docs need to be generated. All docs must be
re-generated if there are any changes to the source ts files.

This removes the in-band generation of docs. The workflow is now:

1. docusaurus api-extractor:init (done once)
2. docusaurus api-extractor:run (done in CI on every commit or in dev
   for dry-runs)
3. docusaurus build