Generate Sidebar Module
This PR does a couple things to enable the generation of an
`api-sidebar` module. Firstly, because docusaurus executes plugins
concurrently "fur da perf" we can't actually use the `loadContent` hook
in the plugin anymore if we want to generate a this file as a side
effect. The reason is that docusaurus-plugin-content-docs is going to
try and load `sidebar.js` before we are done producing `api-sidebar.js`
and thus it will fail trying to require the file.

To solve this issue we must run `docusaurus api-extractor:run ...`
before `build`, `deploy` or `start`. We could come up with some hacky
way to try and sequence these 2 plugins but it is likely to cause other
issues in the future.

Secondly, the module we generate is a module and not a JSON file because
we have no way of knowing the docRoot within the MarkdownFeature and
thus we cannot fully produce the `id` for the sidebar. Instead you must
call the exported function with the relative path for the docRoot if
you're using something other than the default.

That said this turned out really nice. Thanks @scalvert for the
suggestion.