bugfix: Fix issues with Heritage Types Markdown
This fixes an issue where heritage types e.g. Extends, Implements,
References, and how the were printed. For some reason Docusuarus
would not handle the following correctly:

```md
<b>Extends:</b> [SomeBase](./project.some.md)
```

This PR fixes this by specifically coverting `<b>` into `**`
around heritage types.

During this adventure I ran into an issue where the CJS version
of `remark` has a bug which will always escape `_`. Since
api-documenter produces files with `_` in the path this breaks
the escaping breaks links. This is fixed in `remark` 14 however
it is an ESM module.