Add GitHub-flavored markdown support
Allows markdown generator to be able to reason about tables, which isn't
a part of the standard Markdown spec. Resolves an issue where escaped
characters in a table would be removed, creating invalid table output.
(This is needed, for instance when a TypeScript union type is listed in
a table.)