Markdown Back end (#27)
The purpose of this PR is to move away from @microsoft/api-documenter for generating the markdown files. The reasons for this are:

Produce docusaurus markdown in a single pass. Currently we re-parse and repair to make it work.
Support for custom tsdoc annotations like @frameworkItemType which requires parsing
Liberty to parallelize doc generation via worker_threads in the future
Liberty to add different types of sections via annotations in the future